CHORE: entity 클래스 repository패키지로 이동
